pyproject.toml
[project]
name = "annotator-kappa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "pynrrd>=1.1.3",
    "scikit-learn>=1.7.2",
]
